Move dashedBorder into Drawings file
